Add files via upload
Hi all,
Are you a passionate fullstack developer who wants to join a Innovative AI company of the future?
 
“We believe in giving our employees freedom, responsibility, and opportunities to grow!”  

About the company
Wavr Tech AB is leading the development of next level proactive health and safety services. Do you want to join a startup company where you will have the opportunity to influence and impact the journey ahead? Right now, we are expanding our development team in Gothenburg, and are looking for a C# .NET fullstack Software Developer. You will be part of a development team with complete ownership for a number of different systems: customer and internal web systems, DevOps, tools and APIs. We believe in building a lot of our own tools and software in a low level fashion. 

Job description and responsibilities
As part of the team your development skills will strengthen our ability to develop and deliver new and better products to our customers. You are naturally interested in modern technologies with a focus on the Microsoft stack even though we welcome your broad technical knowledge and experience. We follow agile principles and being self-driven is a must, but we naturally think that communication and collaboration within the team is vital for us to grow together. We enjoy an open, helpful and cheerful atmosphere. The tasks will include but not be limited to, design, code, code review, test, document and operate features using technologies like C# .NET, .NET Core, ASP.NET, Blazor, Service Fabric, ZeroMQ, Microsoft Sql Server, Git, Windows, Azure DevOps, Azure CLI, and more.

Who are you?
You will have a Computer Science degree or something equivalent education-wise
Learn fast and are eager to take on challenges and like to share your skills and teach your colleagues
Take responsibility of and own the products you help to deliver

Qualifications
Skilled C# .NET, .NET Core developer

Skilled with multiple databases

Experience designing and building scalable high-availability systems

Experience with test automation

Experience with continuous integration and continuous delivery

Experience with various web technologies

Fluent either in swedish or english

Meriting

Azure infrastructure experience.

Experience with Infrastructure automation tools.

Service Fabric infrastructure and development knowledge.

Machine learning frameworks (e.g Tensorflow, Pytorch, ML.NET)
 

Reason to join us
We are a company where you can influence how we work and be a part of the decision making in how to move forward. Our team is friendly and humble, we lift each other and teach each other. You’ll have the possibility to work remotely and we have flexible hours.


Does this sound interesting?
Selections and interviews will take place on an ongoing basis so do not wait with your application!

NXT Interim and NXT Rekrytering is a niche and innovative company active in the consulting and recruitment industry.
We are active in HR, IT, Finance and Marketing& Sales. We offer packaged services that the traditional companies do not offer today. We have over 20 years of experience in the industry and 3000 unique contacts to help you. 
Our goal is to go from being an innovative challenger in the industry to the company in the industry that everyone wants to beat, contact us today and we will do everything to help you. We have offices in Stockholm, Gothenburg, Malmö, Linköping and Kalmar

How do I proceed
If you are interested, take a first step in showing your interest in this ad. We process the applications on an ongoing basis so do not hesitate but apply today! If you have any questions, please contact Gig Manager Jeanette Ångman on phone: 0704 98 75 56 or email:jeanette.angman@nxtinterim.se